created shared module that exports all common modules split out shared componenets and services moved patient registration to patient project
