+++ frontend/TvCv19.Frontend/ClientApp/projects/patient/src/lib/patient.module.ts
@@ -1,13 +1,22 @@
 import { NgModule } from '@angular/core';
 import { PatientComponent } from './patient/patient.component';
 import { RouterModule } from '@angular/router';
+import { SharedModule } from 'src/app/shared/shared.module';
+import { PatientRegistrationComponent } from './patient/patient-registration/patient-registration.component';
+import { AssignCareGiverComponent } from './patient/patient-registration/assign-caregiver.component';
+import { CameraSetupComponent } from './patient/patient-registration/camera-setup.component';
 
 
 
 @NgModule({
-  declarations: [PatientComponent],
-  imports: [RouterModule
+  declarations: [
+    PatientComponent,
+    PatientRegistrationComponent,
+    AssignCareGiverComponent,
+    CameraSetupComponent,
   ],
-  exports: [PatientComponent]
+  imports: [RouterModule,SharedModule
+  ],
+  exports: []
 })
 export class PatientModule { }
