+++ frontend/TvCv19.Frontend/ClientApp/projects/patient/src/lib/patient/patient-registration/camera-setup.component.ts
