+++ frontend/TvCv19.Frontend/ClientApp/projects/patient/src/lib/patient/patient-registration/patient-registration.component.scss
