+++ frontend/TvCv19.Frontend/ClientApp/projects/patient/src/lib/patient/patient.component.html
@@ -1,2 +1,2 @@
 <h1>it works</h1>
-<!-- <lib-video-receiver fxFill fxFlex [patientID]="patientID"></lib-video-receiver> -->
+<lib-video-receiver fxFill fxFlex [patientID]="patientId"></lib-video-receiver>
