+++ frontend/TvCv19.Frontend/ClientApp/projects/physician/src/lib/patient-list/patient-list.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { FirstLinePatientRouteDataModel } from '../physician-heirachy/firstlinePatientRouteData.model';
-import { PatientModel } from 'src/app/patient-registration/patient-model';
-import { PhysicianModel } from 'src/app/physician-admin/physician-model';
+import { PatientModel } from 'src/app/shared/models/patient-model';
+import { PhysicianModel } from 'src/app/shared/models/physician-model';
 
 @Component({
   selector: 'lib-patient-list',
