+++ frontend/TvCv19.Frontend/ClientApp/projects/physician/src/lib/physician-heirachy/physician-heirachy-route-resolver.service.ts
@@ -7,7 +7,7 @@ import {
 import { Observable, of, EMPTY }  from 'rxjs';
 import { mergeMap }         from 'rxjs/operators';
 import { FirstLinePatientRouteDataModel } from './firstlinePatientRouteData.model';
-import { PatientService } from 'src/app/patient.service';
+import { PatientService } from 'src/app/shared/services/patient.service';
 
 @Injectable({
   providedIn: 'root'
