+++ frontend/TvCv19.Frontend/ClientApp/projects/physician/src/lib/physician-heirachy/physician-heirachy.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
-import { PatientModel, AdmissionStatus } from '../../../../../src/app/patient-registration/patient-model';
+import { PatientModel, AdmissionStatus } from '../../../../../src/app/shared/models/patient-model';
 import { ActivatedRoute } from '@angular/router';
-import { PatientService } from '../../../../../src/app/patient.service';
-import { PhysicianModel, HierarchyLevel } from 'src/app/physician-admin/physician-model';
+import { PatientService } from '../../../../../src/app/shared/services/patient.service';
+import { PhysicianModel, HierarchyLevel } from 'src/app/shared/models/physician-model';
 import { FirstLinePatientRouteDataModel } from '../physician-heirachy/firstlinePatientRouteData.model';
 
 @Component({
