+++ frontend/TvCv19.Frontend/ClientApp/src/app/admin/physician-admin.component.scss
