+++ frontend/TvCv19.Frontend/ClientApp/src/app/app-routing.module.ts
@@ -2,18 +2,18 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { PhysicianComponent } from 'projects/physician/src/public-api';
 
-import { PatientRegistrationComponent } from './patient-registration/patient-registration.component'
-import { AssignCareGiverRouteResolverService } from './patient-registration/assign-caregiver-route-resolver.service';
-import { AssignCareGiverComponent } from './patient-registration/assign-caregiver.component';
-import { CameraSetupComponent } from './patient-registration/camera-setup.component';
-import { physicianAdminComponent } from './physician-admin/physician-admin.component';
+import { PatientRegistrationComponent } from '../../projects/patient/src/lib/patient/patient-registration/patient-registration.component'
+import { AssignCareGiverRouteResolverService } from '../../projects/patient/src/lib/patient/patient-registration/assign-caregiver-route-resolver.service';
+import { AssignCareGiverComponent } from '../../projects/patient/src/lib/patient/patient-registration/assign-caregiver.component';
+import { CameraSetupComponent } from '../../projects/patient/src/lib/patient/patient-registration/camera-setup.component';
+import { physicianAdminComponent } from './admin/physician-admin.component';
 import { PatientComponent } from 'projects/patient/src/public-api';
 
 const routes: Routes = [
   { path: 'physician/:physician-id', component: PhysicianComponent, loadChildren: 'projects/physician/src/public-api#PhysicianModule' },
-  { path: 'physicians/admin', component: physicianAdminComponent },
+  { path: 'admin/registration', component: physicianAdminComponent },
   // Patient workflow
-  {path: 'patient/:id', component: PatientComponent},
+  {path: 'patient-detail/:id', component: PatientComponent},
   { path: 'patient/registration', component: PatientRegistrationComponent },
   { path: 'patient/registration/camera-setup', component: CameraSetupComponent },
   {
