--- frontend/TvCv19.Frontend/ClientApp/src/app/physician.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { PhysicianService } from './physician.service';
-
-describe('PatientAdminServiceService', () => {
-  let service: PhysicianService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(PhysicianService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
