+++ frontend/TvCv19.Frontend/ClientApp/src/app/shared/qr/qr-scanner.component.spec.ts
