+++ frontend/TvCv19.Frontend/ClientApp/src/app/shared/services/patient.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { HttpHeaders } from '@angular/common/http';
-import { PatientModel } from './patient-registration/patient-model';
+import { PatientModel } from '../models/patient-model';
 import { Observable, throwError } from 'rxjs';
 import { catchError, retry } from 'rxjs/operators';
