+++ frontend/TvCv19.Frontend/ClientApp/src/app/shared/services/physician.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { HttpHeaders } from '@angular/common/http';
-import { PhysicianModel } from './physician-admin/physician-model';
+import { PhysicianModel } from '../models/physician-model';
 import { Observable, throwError } from 'rxjs';
 import { catchError, retry } from 'rxjs/operators';
